% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapglview.R
\name{create.mapglview.popups}
\alias{create.mapglview.popups}
\title{create.mapglview.popups}
\usage{
create.mapglview.popups(sfx)
}
\arguments{
\item{sfx}{An \code{sf} object for which to create popups}
}
\description{
Create popups for a mapgl map with all columns except the geometry column.
}
