% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapglview.R
\name{get.breaks}
\alias{get.breaks}
\title{get.breaks}
\usage{
get.breaks(z, breaks.method = c("natural", "quantile", "equal"), n.breaks = 4)
}
\arguments{
\item{z}{A continuous vector from which to generate breaks}

\item{breaks.method}{How to create breaks in legend & interpolation if \code{zcol}
is specified and is a continuous variable. One of \code{natural}, \code{quantile}, or
\code{equal}. \code{natural} uses kmeans clustering.}

\item{n.breaks}{The number of breaks to include in the legend, for continuous
variables.}
}
\description{
A flexible function to internally get breaks according to a given style.
}
