% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapglview.R
\name{get.mapgl.interpolator}
\alias{get.mapgl.interpolator}
\title{get.mapgl.interpolator}
\usage{
get.mapgl.interpolator(x, zcol, palette = viridis::viridis(5), ...)
}
\arguments{
\item{x}{An object of class either \code{maplibregl} or \code{sf}. If \code{maplibregl}, it
will add the data specified in \code{y} to the map; if \code{x} is \code{sf}, it will
render a new \code{maplibre} map with \code{x} added to it.}

\item{zcol}{A column name from the data to interpolate colors across.}

\item{palette}{A character vector describing color hex codes from which to
construct the color palette. Viridis by default. The number of colors
provided doesn't have to match the number of breaks produced.}
}
\description{
get.mapgl.interpolator
}
