% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapglview.R
\name{mapglview}
\alias{mapglview}
\title{mapglview}
\usage{
mapglview(
  x,
  y = NULL,
  zcol = NULL,
  base.map.style = mapgl::carto_style("dark-matter"),
  include.legend = T,
  layer.name = NULL,
  alpha = 0.9,
  size = 5,
  before_id = "place_country_2",
  ...
)
}
\arguments{
\item{x}{An object of class either \code{maplibregl} or \code{sf}. If \code{maplibregl}, it
will add the data specified in \code{y} to the map; if \code{x} is \code{sf}, it will
render a new \code{maplibre} map with \code{x} added to it.}

\item{y}{spatial data to add to the map, if \code{x} is already a \code{maplibregl}
object (it's ignored otherwise).}

\item{zcol}{A column name from the data to interpolate colors across.}

\item{base.map.style}{style JSON to pass onto \code{mapgl::maplibre}. Not used if
the function is being used to add data to an existing map.
\code{mapgl::carto_style("dark-matter")}, \code{mapgl::carto_style("positron")}, or
other.}

\item{include.legend}{Boolean to include legend or not; only relevant if
\code{zcol} is also specified.}

\item{layer.name}{Name for layer. Shows up in legend title. If \code{NULL}, the
name of the object being mapped will be used.}

\item{alpha}{Layer opacity}

\item{size}{For lines or circles, linewidth or circle radius, respectively.
Ignored for polygons.}

\item{before_id}{Name of layer to position the new layer ontop of; passed
onto the add layer call.}

\item{...}{
  Arguments passed on to \code{\link[=get.mapglview.breaks]{get.mapglview.breaks}}, \code{\link[=get.mapgl.interpolator]{get.mapgl.interpolator}}
  \describe{
    \item{\code{z}}{A continuous vector from which to generate breaks}
    \item{\code{breaks.method}}{How to create breaks in legend & interpolation if \code{zcol}
is specified and is a continuous variable. One of \code{natural}, \code{quantile}, or
\code{equal}. \code{natural} uses kmeans clustering.}
    \item{\code{n.breaks}}{The number of breaks to include in the legend, for continuous
variables.}
    \item{\code{palette}}{A character vector describing color hex codes from which to
construct the color palette. Viridis by default. The number of colors
provided doesn't have to match the number of breaks produced.}
  }}
}
\description{
Wrapper to quickly make a map using \code{mapgl}
}
